Implementação da tela de album apenas cenário de error

+++ app/src/main/java/com/thiagofr/jsonplaceholder/data/Album.kt
@@ -1,7 +1,7 @@
 package com.thiagofr.jsonplaceholder.data
 
 data class Album(
-    val id: Int,
-    val userId: Int,
+    val id: Long,
+    val userId: Long,
     val title: String
 )
